requirement.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
PyMySQL==0.10.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
